Conflicts:
	lib/init.c
	lib/socket.c

+++ test-tool/0122_read6_invalid.c
@@ -1,16 +1,16 @@
-/* 
+/*
    Copyright (C) 2010 by Ronnie Sahlberg <ronniesahlberg@gmail.com>
-   
+
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
    the Free Software Foundation; either version 2 of the License, or
    (at your option) any later version.
-   
+
    This program is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU General Public License for more details.
-   
+
    You should have received a copy of the GNU General Public License
    along with this program; if not, see <http://www.gnu.org/licenses/>.
 */
@@ -19,11 +19,12 @@
 #include <stdlib.h>
 #include <string.h>
 #include "iscsi.h"
+#include "iscsi-private.h"
 #include "scsi-lowlevel.h"
 #include "iscsi-test.h"
 
 int T0122_read6_invalid(const char *initiator, const char *url, int data_loss _U_, int show_info)
-{ 
+{
 	struct iscsi_context *iscsi;
 	struct scsi_task *task;
 	struct iscsi_data data;
@@ -118,7 +119,8 @@ int T0122_read6_invalid(const char *initiator, const char *url, int data_loss _U
 		scsi_free_scsi_task(task);
 		goto test2;
 	}
-	if (task->residual_status != SCSI_RESIDUAL_OVERFLOW || task->residual != block_size) {
+	if (task->residual_status != SCSI_RESIDUAL_OVERFLOW ||
+	    task->residual != (ssize_t)block_size) {
 	        printf("[FAILED]\n");
 		printf("Read6 returned incorrect residual overflow.\n");
 		ret = -1;
@@ -164,7 +166,8 @@ test2:
 		scsi_free_scsi_task(task);
 		goto test3;
 	}
-	if (task->residual_status != SCSI_RESIDUAL_UNDERFLOW || task->residual != block_size) {
+	if (task->residual_status != SCSI_RESIDUAL_UNDERFLOW ||
+	    task->residual != (ssize_t)block_size) {
 	        printf("[FAILED]\n");
 		printf("Read6 returned incorrect residual underflow.\n");
 		ret = -1;
@@ -207,7 +210,8 @@ test3:
 		scsi_free_scsi_task(task);
 		goto test4;
 	}
-	if (task->residual_status != SCSI_RESIDUAL_OVERFLOW || task->residual != block_size - 200) {
+	if (task->residual_status != SCSI_RESIDUAL_OVERFLOW ||
+	    task->residual != (ssize_t)block_size - 200) {
 	        printf("[FAILED]\n");
 		printf("Read6 returned incorrect residual overflow.\n");
 		ret = -1;
@@ -249,7 +253,8 @@ test4:
 		scsi_free_scsi_task(task);
 		goto test5;
 	}
-	if (task->residual_status != SCSI_RESIDUAL_OVERFLOW || task->residual != block_size) {
+	if (task->residual_status != SCSI_RESIDUAL_OVERFLOW ||
+	    task->residual != (ssize_t)block_size) {
 	        printf("[FAILED]\n");
 		printf("Read6 returned incorrect residual overflow.\n");
 		ret = -1;
